+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+
+- Add `replacements` options to optionally replace tags with text.
+- Changed `removeWhitespace` option to `preserveWhitespace`.
+- Properly handle nested tag exclusion.
+
 # 0.1.0
 
 - Initial release.